+++ src/sdk/PnP.Core/Model/SharePoint/Pages/Internal/RtePageSettings.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace PnP.Core.Model.SharePoint
+{
+    internal sealed class RtePageSettings
+    {
+        [JsonPropertyName("contentVersion")]
+        public int? contentVersion { get; set; }
+    }
+}