+++ src/sdk/CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [unreleased]
+
+### Added
+
+
+### Changed
+
+- Recognize editor component (CK4/CK5) being used in pages and allow developer to act on that `IPage.EditorType` [jansenbe - Bert Jansen]
+- Retain extra page settings (`isSpellCheckEnabled`, `globalRichTextStylingVersion`, `isEmailReady`, `rtePageSettings` and `htmlAttributes`) [jansenbe - Bert Jansen]
+
 ## [1.9.0]
 
 ### Added